chore: Allows running of tests with extras and tox
simply run tox in the top level directory to run all of the pytest unit
tests in a virtualised environment. I had to install some additonal
extras in order to get the tests to run so I updated the extras_require
dictionary and formatted it so that the [test] flag may be used on
install to provide the needed packages. tox makes this easy with an
extras field in [testenv] so the extra packages are also added to the
virtual environment before pytest is run.