feat: implements driver for the Cryomagnetics 4G
The cryomagnetics 4g power supply is a dual channel power supply with
a persistent heater available on one channel. This driver allows the
user to sweep to a desired field strength. In The user does not need to
manually sweep up to the current circulating through the magnet, activate
the persistent heater and then sweep to the desired set point. The driver
automatically handles this.

Note that the authors claim no responsibility for any magnet quenches that
occur with the use of this driver. Use at your own risk and read the source
code before using.